requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.4
groq==0.15.0
streamlit==1.41.1
aiosqlite==0.20.0
python-dotenv==1.0.1
rich==13.9.4